Move webhook handling to helper
